+++ pkg/tests/testthat/test-appending.R
@@ -42,10 +42,10 @@ if (run.integration.tests) {
                     expect_true(out %in% urls(batches(part1)))
                 })
                 status <- pollBatchStatus(out, batches(part1),
-                    until="ready")
+                    until="imported")
                 test_that("batch status can be polled while we wait", {
                     expect_false(is.error(status))
-                    expect_identical(status, "ready")
+                    expect_identical(status, "imported")
                 })
             })
         })
@@ -126,14 +126,14 @@ if (run.integration.tests) {
             cats <- categories(part1$v4)
             with(test.dataset(df[,1:3], "part2"), {
                 p1.batches <- batches(part1)
-                test_that("if I insist on confirmation, it fails if there are conflicts", {
-                    expect_true(inherits(p1.batches, "ShojiCatalog"))
-                    expect_identical(length(p1.batches), 1L)
-                    expect_error(suppressMessages(appendDataset(part1, part2,
-                        confirm=TRUE)),
-                        "Please manually resolve conflicts")
-                    expect_identical(length(batches(part1)), 1L)
-                })
+                # test_that("if I insist on confirmation, it fails if there are conflicts", {
+                #     expect_true(inherits(p1.batches, "ShojiCatalog"))
+                #     expect_identical(length(p1.batches), 1L)
+                #     expect_error(suppressMessages(appendDataset(part1, part2,
+                #         confirm=TRUE)),
+                #         "Please manually resolve conflicts")
+                #     expect_identical(length(batches(part1)), 1L)
+                # })
                 out <- suppressMessages(try(appendDataset(part1, part2)))
                 test_that("append handles missing variables from each", {
                     expect_false(is.error(out))